+++ CHANGELOG.md
@@ -1,3 +1,11 @@
-## 0.0.1
+# Changelog
 
-* TODO: Describe initial release.
+## [0.0.1] - 2024-04-30
+
+### Added
+- Initial release of Coaster package
+- Implementation of The Elm Architecture (TEA) pattern
+- Core components:
+  - Model-View-Update architecture
+  - Command pattern for side effects
+  - Subscription system